backend/requirements.txt
tensorflow==2.15.0
numpy==1.23.5
Pillow==9.5.0
Flask==2.3.2
flask-cors==4.0.0
flask-limiter==3.5.0
matplotlib==3.7.1
scipy==1.11.3
python-dotenv==1.0.0
google-generativeai==0.3.2 